make `chisel` library compatible
