+++ crates/chisel/src/lib.rs
@@ -5,10 +5,12 @@
 #[macro_use]
 extern crate foundry_common;
 
+pub mod args;
 pub mod cmd;
 pub mod dispatcher;
 pub mod executor;
 pub mod history;
+pub mod opts;
 pub mod runner;
 pub mod session;
 pub mod session_source;
